conf: move MultiSrcConf into lisa/conf.py
Split MultiSrcConf and its friends in their own module, since it has
become a chunky set of classes.
